chatgpt fix
The issue with this code is that every time the 'btn5' or 'btn4' buttons are clicked, a new keypress event listener is added to the 'kanjiInput' element. This means that every time the button is clicked, there will be another listener active, even if a previous one is still running. This will cause multiple event listeners to be triggered when the 'Enter' key is pressed, which is likely not desired. To fix this, the keypress event listener should be added only once, for example, outside of the newKanji() function.